Ensure that the ServiceAccount is created before init Job
On an initial helm install, the database boundary-init Job will run on
a pre-install hook, meaning it tries to apply before the Resources are
created. Since it requires the ServiceAccount which itself is a Resource
this creates a chicken and egg situation.

This commit fixes that issue by also applying the ServiceAccount as a
pre-install hook before the boundary-init Job is created